+++ packages/dust_db_sqlite3/CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
-## [0.1.5]
+## [0.2.0] - 2026-09-07
 
 ### Performance